Implemented fixed tilemaps
I may have to clean the code a bit because it ain't super sexy right now,
but I'm quite happy about that change.
Basically, instead of doing a double for loop to render a TileMap,
I create an empty texture and render my tilemap on the new texture.
Then whenever I wanna display my map I just RenderCopy that texture instead.
So for a 10 by 8 TileMap, I only call RenderCopy once instead of 80 times.
I'll have to see if it's as big an improvement as I think it is.
But for now I'm happy with it.